importer,crosscluster: move type resolver to a different package
Since we're using this type resolver for more than just IMPORT, it makes
more sense to put it in a more fitting package.

Release note: None
