+++ pkg/ccl/crosscluster/BUILD.bazel
@@ -4,6 +4,7 @@ go_library(
     name = "crosscluster",
     srcs = [
         "addresses.go",
+        "crosscluster_type_resolver.go",
         "errors.go",
         "event.go",
         "settings.go",
@@ -15,5 +16,15 @@ go_library(
         "//pkg/repstream/streampb",
         "//pkg/roachpb",
         "//pkg/settings",
+        "//pkg/sql/catalog",
+        "//pkg/sql/catalog/descpb",
+        "//pkg/sql/catalog/typedesc",
+        "//pkg/sql/pgwire/pgcode",
+        "//pkg/sql/pgwire/pgerror",
+        "//pkg/sql/sem/tree",
+        "//pkg/sql/sqlerrors",
+        "//pkg/sql/types",
+        "@com_github_cockroachdb_errors//:errors",
+        "@com_github_lib_pq//oid",
     ],
 )
